python-agent-framework/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
streamlit>=1.29.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.18.0
scipy>=1.11.0
supabase>=2.0.0
pydantic>=2.5.0
typing-extensions>=4.8.0
